Use long for ID parameters in guest pre-order
Change several ID types from int to long to support larger identifier ranges and ensure consistency between models and logic. Updated GuestChoiceModel constructor (MenuItemId, GuestId) and PreOrderLogic methods MakeGuest, GetGuest, InsertGuestChoices, and GetGuestById to accept long reservation/guest/menu IDs. No behavioral changes besides type widening.

+++ ReservationSysteem/DataModels/GuestChoiceModel.cs
@@ -9,7 +9,7 @@ public class GuestChoiceModel
     {
     }
 
-    public GuestChoiceModel(int MenuItemId, int GuestId, int Quantity)
+    public GuestChoiceModel(long MenuItemId, long GuestId, int Quantity)
     {
         this.MenuItemId = MenuItemId;
         this.GuestId = GuestId;
